Introduce conditional variants for `ApplyMore`
The org.axonframework.modelling.command.ApplyMore interface provides the two methods ...#andThenApply and ...#andThen which are used e.g. in the AggregateLifecycle to apply a payload or execute some code in the correct order.

The conditional variants of these two methods provided in this commit make it easier to conditionally apply a payload or execute code in a more fluent api (see example below).