Add coverage for `ApplyMore#andThenIf`
Adjusting the AnnotatedAggregateTest to cover `ApplyMore#andThenIf` as well by using a resource injected as command handling parameter. To allow injecting a stub resource as command handling parameter, the StubRepository builder has been adjusted. This required also to slightly alter the setup for the regression test related to #1506 to fit the new command handler of AggregateRoot.