EventProcessingModule should lazily initialize processors
The eager initialization of event processors can cause circular dependencies when one of the event handlers relies on the configuration. This is espectially the case when using an external dependency injection mechanism to wire some beans.

Event Processors are initialized at the latest when the configuration starts, or earlier when the processors are retrieved from the configuration.