Ensure the context is maintained between upcaster invocations in single stream
This commit fixes an issue where a new context was created for each individual event. This caused a context aware upcaster to lose any contextual information while upcasting a single stream of events when loading events for an aggregate.

This commit also fixes an issue where snapshots would be considered when loading an aggregate from its initial position, even when an explicit attempt to load snapshots had already been performed.