React differently to primitive objects in the EqualFieldsMatcher
JDK17 does not allow reflective access to java.lang class, like the
primitives. Due to this, if a user uses a primitive type in commands,
events, queries, deadlines, etc, the EqualFieldsMatcher will throw an
exception. Ideally, the EqualFieldsMatcher wouldn't be exposed through
the Matchers utility directly, since that would allow us to attach an
equals check before proceeding with the deep equals that's currently
performed. However, the ResultValidatorImpl and CommandValidator rely on
 it at the moment. As such, the EqualFieldsMatcher should check if it
 deals with a primitive type and not proceed with the reflective access
 if this is the case.

#2189