Use equals() first for aggregate state check
This avoids performing reflection on classes that may not be declared
open to the module that contains the fixtures. Typically, these classes
would implement a suitable equals method.

This change will test equality using equals first, before attempting to
check fields. If equals returns false, it will still check whether an
equals method is explicitly implemented. Only when that is not the case,
a field-to-field comparison is performed using the same approach for
each individual field.