Adjust usage of the EqualFieldsMatcher
A new implementation should be used instead of the EqualFieldsMatcher.
This new implementation should firstly check if equals returns true. If
it does not, we should check whether equals is overridden by the type
being matched. If it isn't, we fall back to field equality, with a
try-catch(Exception) around it. In case of a caught exception, we log
the exception on warn level and assume the types do not match. This new
matcher should be exposed through the Matchers utility and used by the
framework instead of the deprecated equalTo matcher.

#enhancement/field-matcher