Fine tune validation steps
Instead of explicitlyUnequal(), check for null and whether either class
has an equals method through or operations. As this is proceeded by
Objects#equals, the scenario of (1) either being null or (2) either not
having equals implemented points out equals failed and thus the state
check should fail. Next to this, adjust indentation for clarity.

#2196