Make GenericDeadLetterEntry package-private and add id
The GenericDeadLetterEntry can just as well be used by future other
implementations of the DeadLetterQueue. Furthermore, we can rename it
from DeadLetterEntry to DeadLetter, as that describes the intent
perfectly fine. This change requires validation and adjustment of _any_
occurrences of
"entry" in methods, JavaDoc, comments, and variables. Lastly, the
DeadLetter can benefit from an 'identifier' value and getter for
non-in-memory versions of the DLQ.

#2021