Check whether a WorkerLauncher is active before starting another one
Before the TrackingSegmentWorker may start a WorkerLauncher, it should
first validate whether a WorkerLauncher is active at that moment in time
Although the overall window of opportunity is tiny, there's an option
that two WorkerLaunchers can concurrently start a TrackingSegmentWorker
for a given segment. This, in turn, means events may be handled twice.
How a TrackingEventProcessor could reach this scenario is not clear at
this stage, sadly enough.

#2232