Adjust release operations
Instead of filtering on the entire DeadLetter when releasing, we should
use the QueueIdentifier. This approach better matches how letters are
stored and will make RDBMS implementations of the DQL more efficient in
releasing too, as the index will most likely reside on the
QueueIdentifier. As an added benefit, introduce the release(String
group) operation, clearly tailoring towards the
DeadLetteringEventHandlerInvoker that needs to release based on the
processingGroup.
Lastly, remove any mention of entries in interfaces and the in-mem
implementation, as that might suggest JPA.
 sug

#2021