Check whether the identifier is present during prepareCommit
During the commit phase of an Aggregate, the identifier is expected to
be present. Otherwise, no events or state-stored aggregate can be
inserted. However, if the identifier is not present at this point that
means the command handler decided not to create the aggregate. As such,
no commit action is necessary. Validate this behavior with a test case
targeted towards a command handling constructor without publishing any
events.

#2242