Configure Cache and LockFactory based on beans
First, add a test case to validate whether the configurable bean names
are taken into account. Secondly, add a test case that defines a
Repository bean name, as this overrides configuration of the
CommandTargetResolver, Cache, LockFactory, and SnapshotTriggerDefinition
Follow this up by any fixes to ensure the components are configured as
expected.

#bug/aggregate-stereotype-cache-config