Purge closed units of work
Purge closed units of work when CurrentUnitOfWork.isStarted and
CurrentUnitOfWork.get are invoked. In either scenario there's a
possibility that a closed unit of work is still present while another
process tries to attach a hook in one of the phases. By purging already
closed entries we'll solve this predicament.

#bug/clear-closed-unit-of-works