Ensure all dispatchable messages are serialiable by Jackson and Xstream.
Adds a testcase for each message type that checks whether it can be serialized, and whether the contents are still the same after serializing and then deserializing it.

Useful for occurrences when you want to serialize entire messages instead of only the payload. For example, this will be very useful in the upcoming DLQ feature of Axon 4.6.0