Fix duplicate command handler detection.
This fix is achieved by registering for each signature instead of each Executable. The signature is a combination of the method's name and its parameters. Different Executables can have the same signature, for example in polymorphic aggregates. When the signature is the same, we only want to register once, since this is not a true duplicate handler. At runtime, the AnnotatedHandlerInspector will resolve the right Executable based on the aggregate type and the command.

Fixes #2243