Rethink the DeadLetteringEventHandlerInvoker
The DeadLetteringEventHandlerInvoker should align with the new changes
around the EnqueuePolicy, EnqueueDecision, and the
SequencedDeadLetterQueue. As such, the invoker should construct the
GenericDeadLetter to enqueue. But not before it validates the
EnqueuePolicy taking its EnqueueDecision into account whether to
enqueue to begin with.

As we'll not provide any automatic retries anymore, the
DeadLetteringEventHandlerInvoker should provide a means to trigger a
retry for its processing group. As the API for the invoker is not
finalized yet around this topic, some todos are left. A form of retry
method will at the very least invoke the DLQ#process operation. This
means the EvaluationTask should align with the fact it will handle a
single dead-letter instead of a while-available.

The task will thus receive a single letter, handle it, and construct
an EnqueueDecision based on handling. On success, it'll share the
decision to evict() the letter. On failure, it'll invoke the
configured EnqueuePolicy. To clarify the intent further,
the EvaluationTask is renamed to DeadLetteredEventProcessingTask.

#2021