Remove the SequenceIdentifier and add markTouched()
Remove the SequenceIdentifier as we enforce construction of a DLQ per
Processing Group. Due to this, the addition of "group" on the
"sequenceIdentifier" to differentiate between sequences is no longer
necessary. Similarly, remove the DeadLetter#identifier method, as the
letter can be uniquely identifier based on it's state already. If
concrete implementations need an identifier, they can add it there.
Lastly, add DeadLetter#markTouched to signal a sole adjustment of the
touched field.

#2021