Revamp the SequencedDeadLetterQueue
- Change the generic from an implementation of DeadLetter to an
implementation of Message. Doing so provides more certainty on the
maintained letter instances.
- Remove the idea of the SequenceIdentifier entirely, as we cannot
guarantee we can reconstruct the identifier. This follows from the
argument that the actual "sequenceIdentifier" is an Object the user can
choose. Thus exposing this on the interface shows a promise that we can
reproduce the identifier, something we cannot promise. This change
merits the introducing of "sequenceIdentifier" on some methods, and
replacing the "SequenceIdentifier" class for "Object" on others.
- Make requeue take in a letter update function instead of a throwable.
This more clearly aligns with the flexibility to update the letter once
the original letter is found in the queue.
- Adjust the DLQ#process operation such that it will process an entire
sequence on its own for as long as there are letters and the processes
decided to evict it. Doing so means we keep ourselves to the certainty
that a given sequence is only handled by a single process, thus guarding
 against concurrent access and adhering to the sequence ordering.
- Remove any notion of "group" on the DLQ. This is in line with removal
of the SequenceIdentifier. We will thus expect the end user to use a DLQ
 per group themselves, ensures different processingGroup results do not
 mix with one another.

#2021