Revamp the DeadLetteringEventHandlerInvoker
- Adjust all the generics according to the changes in the
SequencedDeadLetterQueue and EnqueueDecision. These now expect an
implementation of Message i.o. one of DeadLetter.
- Remove the ListenerInvocationErrorHandler. It incorrectly gives
control to the user to no longer dead-letter anything, which does not
make sense when using the DeadLetteringEventHandlerInvoker.
- Remove the group and processingGroup fields. These are no longer
needed to filter the sequences to process, as a DLQ will be configured
per processingGroup.
- Adjust process to invoke the DLQ just once. The ownership of looping
through a sequence now belongs to the DLQ instead of the
SequencedDeadLetterProcessor, as that way we can protect concurrent
access better.

#2021