Add synchronized blocks around the deadLetters collection
To further protect the InMemorySequencedDeadLetterQueue against
concurrent changes on the dead-letters collection, we should add some
synchronization around it when accessing. Main point of attention here
are the evict and requeue operations which might remove a sequence or
bring it back. Due to this, contains and enqueue should also synchronize
The process operation does not receive any synchronized-blocks as it
works through the takenSequences set.

#2021