Make given-phase error handling configurable
Currently, an exception thrown in the given-phase of a SagaTestFixture
cause the fixture to proceed to the when-phase. This may have undesired
side effects, as exception in the given-phase aren't clearly shared with
 the end-user. To that end, it would be better to not invoke the
 (optionally) configured ListenerInvocationErrorHandler if the
 RecordingListenerInvocationErrorHandler has not started yet.
If it is started, the ListenerInvocationErrorHandler should be invoked.
Make this setting configurable directly through the FixtureConfiguration
 interface. This allows users to proceed or fail on given-phase
 exception.

#enhancement/sage-fixture-error-handling-in-given