Add test case validating the issue
Add a test case to validate the annotated command interceptor
predicament. For clarity, the issue is when an Aggregate has an
Entity/AggregateMember. The entity has a command handler, for which the
aggregate root has a @CommandHandlerInterceptor annotated method. If
this method uses the InterceptorChain, it should be invoked regardless
of the existence of the Entity yes/no.

#2393