Move retrieval of suitable handlers to actual invocation
We, apparently, allowed the invocation of CommandHandlerInterceptor
annotated methods for commands handled by an entity without any concrete
 instance. To keep supporting this, we need to move the retrieval of the
 suitable handler, where we check whether we can forward a command to
 an actual instance, to right before the handler is invoked by the
 interceptor chain.
Furthermore, we can drop the if/else branch based on the existence of
interceptors, as the DefaultInterceptorChain already makes this check.
Lastly, drop the method-level warnings suppression in favor of per line.
This will make warnings on new code sections inside the methods more
evident.

#2393