Move Javadoc plugin to javadoc profile, activated on performRelease
Move Javadoc plugin to javadoc profile, activated on performRelease.
Doing so, we ensure the automatic JDK11 build does not run the JavaDoc
operation. This operation currently fails on being incapable of finding
the Nonnull / Nullable annotation during documentation construction. My
hunch has to do with the used namespace. Instead of figuring this out
right now, we can deactivate the unnecessary construction through the
use of this profile.