Fix regression for GenericJpaRepository autoconfig
When annotating an aggregate with `@Entity`, the autoconfiguration
should configure a GenericJpaRepository, instead of an
EventSourcingRepository. In 4.6.0 and 4.6.1, this was not the case
anymore.
This commmit re-introduces the autoconfiguration of the
GenericJpaRepository.
Additionaly, configuring a repository based on the naming convention
`aggregateName+"Repository"` resulted in a circular dependency due to a
bug. This is also resolved in this commit.