Update design principles according to 13-10-2022 session
Update the design principles according to the 13-10-2022 session,
between Allard, Mitchell, Gerard, Yvonne, Sara, and Milan. During this
session, we covered in greater detail the configuration API. Currently,
users need to provide components that they should not necessarily touch,
 ever. Finding a means to abstract that away from them, with potentially
 higher level configuration components and a subsequent decorator
 pattern internally on our end, may help in this perspective. We also
 discussed whether the ServiceLoader should stay, that Serializers
 should be configured on the buses, and that exception should not be
 thrown from within functional code (i.e., use the Mono/Flux monad API).
The last topic revolved around managing the registration/setup of
message handlers and message dispatcher, and how we're going to release
the current hard requirement on concrete Java classes towards name
(space)s, revisions, concrete types and schemas. There's more to be
covered on this topic in the following session.

#axon-5-design