Added Spring Resource Injector to Application Context
There is a regression in the Spring Autoconfiguration since 4.6, causing the Resource Injector to not be available as a bean in the application context. As a result, the Resource Injector used by default is the one taking resources from the Configuration, rather than using Spring to autowire dependencies. In most cases, this will behave the same way. Only when using Qualifiers or spring-specific dependency injection options, this may lead to issues.

This commit will enforce the same behavior as 4.5, where a user-defined resource injector takes precedence, while defaulting to the SpringResourceInjector.