Reintroduce overwritten commits from issue #1905
The changes introduced by user dmurat in PR #1905 where accidentally
overwritten by a force push. As the master branch no longer contained
the changes, nor did the original branch from dmurat still exist, we
will copy-paste the adjustments by downloading the master branch at the
point of merging PR #1905. This, sadly, causes the commits
5b526fe28d1b32986cef473861bdd57cacef4882 and
191a326dbac4f696e0cb7588cdb2e3bfed304e69 (which thus aren't part of the
master branch) to not belong to dmurat. To spot the original commit, it
is recommended to check out this page: https://github
.com/AxonFramework/AxonFramework/pull/1905/commits

#2454