Adjust behavior around EventUpcaster configuration.
When implementing the new Spring configuration approach for AF 4.6.0, we
 adjusted how the DefaultConfigurer constructs the default
 EventUpcasterChain. It would now look for the "default" component
 first, which in the DefaultConfigurer doesn't do anything. IN the
 SpringConfigurer, this method assumes a single, or otherwise primary,
 bean to take that roll. Although this suffices for the majority of
Axon's infra components, the EventUpcasterChain *is* an EventUpcaster
itself. This allows chaining these chains together, so that users can
group their upcaster chains if desired. To fix the predicament, the
original default EventUpcasterChain construction should return to the
DefaultConfigurer. Hence, all registered EventUpcasters are combined
into a single EventUpcasterChain.

#2445