Adhere to the @Order annotation on message handling components
Event Handling Components should adhere to the presence of the @Order
annotation when in a Spring environment. This allows users to impose an
ordering on their event handling classes within a given Event Processor.
We can follow the approach taken in the deprecated
SpringAxonAutoConfigurer for upcaster beans, which sorted bean
references based on the value of the @Order annotation through an
implementation of the ListableBeanFactory.

#2449