Adjust log level from WARN to DEBUG
The method used to retrieve the bean names,
ListableBeanFactory#getBeanNamesForType(Class), will *always* return
more than one entry if there are several beans of the given type or a
subtype. The only reason those beans won't show, is if they're not
annotated with @Aggregate. And if they aren't, the polymorphic
configuration wouldn't be polymorphic anyhow. As such, marking this log
statement as warn was unnecessary to begin with.

Resolves #2512