Fine tune test cases for resiliency
- Move common used fields to constants
- Remove manual cache clearing, as that's a job of the IT
implementations to define when these are cleared.
- Maintain references to the associations set, to ensure the weak
reference implementation doesn't fail on the missing associations
- Validate whether the saga cache entry is present before verifying the
contents, instead of an assertion. This it to cover for cache
implementation that may have cleared out the entries between the (bulk)
updates and validation. As this is a perfectly valid scenario, just
don't validate the cache if it's not there instead of breaking the test
case.
- Suppress unimportant warnings

#2517