Adjust solution in EntryListenerValidator
As there's a window of opportunity that a Saga is searched for right
after it's removed, the removed check may currently fail from time to
time. As this deviation isn't important for the test cases, we can get
rid of clearing out the removed set when an entry is added. Added,
clean-up the validator a bit further for clarity.

#2531