Introduce PostgresSagaSqlSchema#sql_loadSaga from #346
Introduce the customized loadSaga(Connection, String) method introduced
in #346 for AF 2.4 by sgrimm (Steven Grimm). This adjustment allows
users to enforce the PostgresSagaSqlSchema to use a row lock when
loading a saga. This allows protection against concurrent access on a
Saga between distinct nodes. This scenario is only expected to occur
when a non-StreamingEventProcessor is used.

Resolves #947