Introduce whenConstructing and whenInvoking
Introduce whenConstructing and whenInvoking to the TestExecutor.
whenConstructing can be used to simulate invocation of an Aggregate
constructor in the test fixture, when users prefer to not use the
CommandHandler annotation in the Aggregate itself AND have not
registered an (Annotated-)Command Handler with the fixture. Add the
whenInvoking method for a similar scenario, but then not for
construction but "regular" state changes.

#1509