Deal with failing RepositoryWiringTest class
The JDK17 build fails on the RepositoryWiringTest, specifically on
wiring several aggregate Repository beans. This has likely to do with
SpringBoot 3, as moving the class its previous package
(axon-spring-boot-autoconfigure) makes it run. As to not pause the
release longer, this class is copied over to its previous package, and
the tests are ignored. This fix also requires adding back the customizer
 to the SpringAggregateLookup class.

#2578