Define a stub AxonServer server to use during testing
Several of the tests require the existence of a connection to what they
expect to be Axon Server. Although we can set up a test container for
this, doing so will lengthen the tests quite some. Especially since the
only required operation to work, is registering a client. Hence, we can
copy the stub solutions present in the axonserver-connector module, and
fine tune them for these tests. To not get clashes in port numbers
between closing server instances, the TcpUtils is also reused, setting a
 system property for the entire test class (and clearing it afterwards).
While add it, adjust some of the tests to use the
ApplicationContextRunner approach, streamlining the test cases.

#enhancement/shorten-build-logs