Fix OpenTelemetry support
The current implementation of OpenTelemetry support is wrong. It confused starting a span (starting its timer) with making it the current span for the thread (so that any span created by this thread gets the current as a parent).

This has various effects, such as spans being nested despite being siblings, some span collectors not working at all, or other various unwanted effects.

To remedy the situation, the SpanScope is introduced in this commit. Before ending a span, all open scopes should have been closed in order to prevent context leaking into the current thread, messing up the traces.