Fix OpenTelemetry support on queries
During the last change for OpenTelemetry on 4.6.x branch, an error was made by wrapping the response processing task in a span that is then executed multiple times. This wrap has been removed, and the span is now started by the task itself when appropriate.

For all makeCurrent calls, we now use try-with-resources statements. That ensures scopes are closed even when exceptions occur while the bus is shutting down or a dispatch interceptor throws an exception.