Add a document discussing reactive-native.md
Add a document containing our combined knowledge about what it means to
be reactive, what the pros/cons are of certain implementations, a
gathering point for useful links and a concrete plan to support both
reactive and non-reactive programming styles for framework users.
Discussed this of the course of two sessions with Gerard, Milan,
Mitchell, Sara, and me, on the 2nd and 3rd of March 2023.

#axon-5-design