Fix OpenTelemetry warning
When a command handler applies an event and stages events for publication in the UoW, but then fails, an error is logged by Axon Framework: Span was attempted to be ended while not started yet

This is because the cleanup phase (that ends the span) is called even when the span is not started.
Since it's nearly impossible to tell what has been executed and what hasn't, the commit span has been split up for each phase in the UnitOfWork - preCommit, commit and postCommit.

Last but not least, the error logs have been reduced to warn - we can handle the issues gracefully.