Describe deadline design principles
Following the discussion on the 7th, where Sara pointed out her
discontent with the current Deadline approach, we went on an extensive
philosophical route to discuss how they should be constructed. We
concluded to not fully agree on everything, but that we do feel the
current API AND implementation should adjust to accompany our
current-day findings.
Discussed this pointer with Allard, Gerard, Milan, Sara,
and myself, on the 9th of March.

#axon-5-design