Remove payloadType tag from EventProcessorLatencyMetric
The latency per payload type makes no sense to track for an event processor. When monitoring this metric, this will make the value dependent on the rate of certain types of events.

For example, when an event only takes place once a week, and is delayed because of an error, this metric will stay at a high value for the remainder of the week