Further clarify the ease-of-use guideline
The ease-of-use guideline can be broken down into the steps to take from
 within the implementation. Firstly, a bare-bone approach for maximum
 freedom should be present. This can be viewed as the current solution
 where a user directly subscribes message handlers to a bus. Secondly a
 declarative configuration style employing lambdas should be provided.
This second step utilizes the provided API in step 1. Furthermore, it
opens up the capability to not use annotations while maintaining the
same power AF4 currently provides. And lastly, an annotation-based easy
solution should be present. As with step 2, this step utilizes its
predecessor.
Discussed this pointer with Allard, Gerard, Sara,
and myself, on the 14th of April.

#axon-5-design