Move inner classes out
The use of inner classes for the Aggregates, especially on an
@Configuration annotated class, causes different wiring behavior than
when Spring inspects non-inner classes. For non-inner classes, any
abstract class within a package (meta-)annotated with @Component is
ignored  (expect when these include @Lookup annotated methods). For
inner classes this validation step is not taken. As such, the root type
of a polymorphic aggregate is not included for regular component
scanning, while it is for the polymorphic aggregate auto-configuration
tests. Hence, we should use regular component scanning in the tests too.

#bug/polymorphic-aggregate-factory-root-type