Expand upon command modelling design principles
* Add "Aggregate" to the title to clarify this section's intent
* Define aggregates need to be configurable through annotations and
through a declarative configuration API. This allows users that wish to
not have framework/library logic in their domain model to use AF
nonetheless.
* Describe that the current AggregateLifecycle#apply obstructs the fact
that an applied event is first handled inside the aggregate before
returning to the command handling method. Although this behavior is
desirable, users should be able to make this explicit in their
application instead of it happening under the hood. Furthermore, we'd
need such an explicit solution to be in place for the declarative
aggregate configuration.
Discussed this pointer with Gerard, Sara, Milan,
and myself, on the 26 of April.

#axon-5-design