Add required fields throughout impl and builder
Add the fields this DLQ will need to implement all the methods, largely
resembling what's present in the in-memory and JPA versions of the
SequencedDeadLetterQueue. Next to these, add the ConnectionProvider and
DeadLetterSchema, combined with a method to construct the basic table.
The latter resembles the approach followed by the JdbcTokenStore and
JdbcEventStorageEngine, resulting in a similar solution throughout for
JDBC users of Axon Framework.

#feature/2350-jdbc-dlq