Optimize deadLetterSequence method by introducing paging
As a dead-letter sequence may contain a lot of events, we need to be
mindful of the query done. As a resolution, we follow the approach
taken in the JPA DLQ, which is to page results. To that end, a paging
iterable should be included to the common/jdbc package. Furthermore, the
 letterSequenceStatement should allow a firstResult and maxSize
 parameter to be included for this.

#2350