Implement evict and construct JdbcDeadLetter
To be able to cleanly evict a dead letter, we want to use the dead
letter identifier is used in the database. To that end, the methods in
the JDBC DLQ should return a Jdbc-specific DeadLetter instead of the
GenericDeadLetter. Furthermore, the factory should provide an
evictStatement operation. Lastly, adjust all mentions of the generic M
for E, as we expect and EventMessage instead of a Message

#2350