Resolve TODO
Resolve TODO as there's no one solution for all databases. We'd require
two result types from the database, namely a value and no value to be
able to differentiate between the start of the sequence or an index
somewhere within the sequence. As we cannot guarantee this is in place
for all databases, defaulting to start from 1 is easiest.

#2350