Adjust the SpringAggregateLookup to set the new aggregate factory
The approach towards defining the root/top aggregate name should change
to accommodate for the fact an abstract @Aggregate annotated root class
does not occur in the ApplicationContext. Due to this, the aggregate
hierarchy is essentially wrong. We should instead default to a
reasonable root aggregate bean name for our own purposes, and provide
this name to the SpringPrototypeAggregateFactory

#2689