Add AxonServerContainerUtils to ease use of the AxonServerContainer
Add AxonServerContainerUtils to ease use of the AxonServerContainer.
With the merge of EE and SE, users are inclined to initialize the
cluster in all cases. Regardless of whether they're running single or
multi node. Although doable, is friendlier to provide utilities for this
Note the utils are a form of the test utils present in the
axonserver-connector-java project

#feature/single-axon-server-test-container