Default to ReplayToken upon creation of new event processor
Users expect processors to initially behave as a replay. So, if the head of the event store is at 10.000, most users expect the `ReplayStatus` flag to be of type `REPLAY`, instead of `REGULAR`.

This PR changes the default configuration to do just that. The initial token has been adjusted to by default be a `ReplayToken`, coming with all the functionality of doing a replay.

Use-cases include event processors that act upon an external non-idempotent system, where you during a replay don't want that action to happen but still want to insert data into the database for later use.