Add isProcessingEvents flag to the WorkPackage, since only then do we...
Add isProcessingEvents flag to the WorkPackage, since only then do we
want to have the Coordinator extend the claim for it. Otherwise, the
WorkPackage should be perfectly capable of doing it itself. This flag
should be set in the processEvents section of the WorkPackage when it
actually has events. Furhtermore, move the transactionManager usages
into the extendClaim operation, as well as put back the
lastClaimExtension value. The latter ensures neither the WorkPackage or
the Coordinator will extend the claim too frequently. To be threadsafe,
adjust the lastClaimExtension into an AtomicLong.

#2785