Rename EventProcessorConfiguration to Eventhandling
Although Spring resolved the name for the properties based on the
ConfigurationProperties annotation on the getter, it is unable to set
the properties in this format. To keep the property path name that's
already propagated throughout our user base, I've opted to rename the
EventProcessorConfiguration to Eventhandling, to align with the path
"axon.axonserver.eventhandling". Although we should strictly regard this
 as a breaking change, I think the chances are minimal somebody is
 directly interacting with the object i.o. through property files.

#bug/load-balancing-strategy-in-property-file