Implement StreamingEventProcessor.claimSegment
In order to provide smoother Event Processor operations in the Framework going forward, we have decided to implement a specific action that claims a token for a processor.

Previously, moving processing from one instance to another could only be done by releasing the segment on all instances but the desired, and awaiting the token claim interval of the desired instance. This led to a prolonged period of not processing of events, which is undesirable.

The action will be handled in the context of the processor and claim the token in the store. Then upon the next WorkerLauncher/Coordinator operation (which is queued at the same time) processing of it will start. This works both for the PSEP and the TEP.