Replace initialNrOfPermits for permits
Replace initialNrOfPermits for permits for ease of use. Especially for
command and query flow, only the permits are used. Using a property
named "initial number of permits" suggests you will run out of permits
eventually, which isn't the case. Thus deprecate all mentions of
initialNrOfPermits and add the fields permits (to which the deprecation
should refer).

#enhancement/clean-up-as-config-and-config-usage