Removed spring-devtools.properties
With the new approach to detection and registration of handlers in the Spring Boot Autoconfiguration module for Axon, there is no need to register jars for the restart server. This registration stood in the way of using test containers in combination with devtools.

(cherry picked from commit cc1ada0fe370c01ccce222462ad3bca61bc5b259)