Deprecate ReflectionUtils operations
Deprecate ReflectionUtils operations that aren't easily supported in
JDK9 and up. As it is unclear at this stage what the effort is to
replace these methods within AF, that work is slated in issue #2901 for
release 5.0.0. All usages of the deprecated methods are suppressed for
the time being to clean the build from warnings

#enhancement/resolve-build-warnings