Provide ThreadFactory builder method
Provide ThreadFactory builder method. This introduction aligns with the
deprecation of the Disruptor constructor, which expects setting a
ThreadFactory instead of the Executor. Along with introducing a
ThreadFactory builder and using the provided ThreadFactory in the
Disruptor(), we should adjust the JavaDoc to clarify where the Executor
and ThreadFactory are used for.

#enhancement/resolve-build-warnings