Resolved apparent memory leak in GapAwareTrackingToken
When using a JPA or JDBC based event store, the tracking token need to track gaps to avoid missing events in uncommitted rows in the table. The creation of tokens in the presence of large gaps was inefficient. This caused a large number of java objects to be created, even though these gaps did not need to be tracked. Deleting events and performing a replay would have the same effect.

This commit changes how gaps are copied between tracking tokens, reducing the number of entries in TreeSets and therefor reducing the number of created objects.