Fix Exception on using OpenTelemetry and ScatterGather queries
The onClose handler on ScatterGather queries calls the `end` operation on Spans a second time. This, however, has been restricted by an exception as this can indicate thread leaks.

However, calling the `Span.end` on both `close` and finishing of advancing the stream is necessary, as users might not always close a stream explicitly. Likewise, the stream may be closed without reading it completely.

I have added an AtomicBoolean as close gate on the closeHandler, so it's always called exactly once.

The TestSpanFactory has been extended with the same behavior as the OpenTelemetrySpan to catch any other instances of this happening during the tests