First steps towards the desired structure of the CommandBus
Introduced several smaller components that take on one of the responsibilities of a command bus, rather than having a single class that can do it all. A builder is introduced to allow easy configuration of these "layers".

Removed and moved some classes to better match the future layout of framework components.

Still requires additional testing for the new components.