Move the AxonServerConfiguration higher and add property-based...
Move the AxonServerConfiguration higher and add property-based
AxonServerConnectionDetails bean if none is present. This shift is
necessary to decouple to direct setting of the routingServers on the
AxonServerConfiguration by the ConnectionDetails, as this override is
otherwise overridden again by a property bean post processor.
Furthermore, adding the property-based connection details bean allows us
 the certainty there will be a connection details bean at all times

#3022