Massive amount of work in progress
Introduces numerous new interfaces concerning event storage, intended to
 not clash with the existing code to thus move the full integration to
 a later stage.
Noteworthy additions are the AsyncEventStore, AsyncEventStorageEngine,
AppendCondition, SourcingCondition, StreamingCondition, and
EventCriteria. The first two reflect the old interfaces, but use the new
 API. The latter four, the conditions and the criteria, reflect our
 intent with DCB. Added, the AsyncEventStore API is already used by the
 AsyncEventSourcingRepository, and it's test (for which it is mocked) to
 validate if such an API would be feasible to use.

Furthermore, There's an implementation of the
AsyncEventStore and AsyncEventStorageEngine, called the
AsyncEmbeddedEventStore and AsyncInMemoryEventStorageEngine. The
AsyncEmbeddedEventStore is mostly a wrapper around the
AsyncInMemoryEventStorageEngine for now. The
AsyncInMemoryEventStorageEngine takes a similar approach to the
InMemoryEventStorageEngine, although it uses MessageStreams and
CompletableFutures instead.

Lastly, I have added an implementation of the AppendEventTransaction,
called QueueingAppendTransaction. This impl. computes if there's an
event queue on the given ProcessingContext. During the computation, it
will add an on-prepare-commit hook to add events to the given
AsyncEventStorageEngine, and when that completes, to add the position of
 the last event to the context to be used as a return value.

Note that there are also several TODOs to cover for parts we need to
look into for this issue or for future issues.

#3101